requirements.txt
python-dotenv==1.0.0
requests==2.31.0
tenacity==8.2.2
pandas==2.0.2
matplotlib==3.7.1
scipy==1.11.0 # Only required for z-score normalization